+++ Convertor/BusinessContent/Selfie/SelfieViewController.swift
@@ -56,8 +56,15 @@ extension SelfieViewController: SelfieViewDelegate {
     }
 
     func makeUp() {
+
         log.message("[\(type(of: self))].\(#function), DarkMode: \(DarkMode.style)")
-        view.layer?.backgroundColor = NSColor.perseusBlue.cgColor
+
+        // view.layer?.backgroundColor = NSColor.perseusBlue.cgColor
+
+        if isHighSierra {
+            view.window?.appearance = DarkModeAgent.DarkModeUserChoice == .on ?
+            DARK_APPEARANCE_DEFAULT_IN_USE : LIGHT_APPEARANCE_DEFAULT_IN_USE
+        }
     }
 
     func localize() {