+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Localization.
 - CPL dependency to v1.6.1.
+- DarkMode appearance.
 
 ## [0.0.2] - [2025-11-19], Developer Release